Was making better progress earlier. Back to more static.

+++ svr/svrServer.cpp
@@ -94,7 +94,7 @@ void svrServer::startWorker(Worker *worker, int nthreads) {
 }
 
 
-void svrServer::setImage(unsigned char *pixels, unsigned int jpegSize) {
+void svrServer::setImage(unsigned char *pixels, long unsigned int jpegSize) {
   unsigned char* p = (unsigned char*) malloc (jpegSize);
   memcpy(p, pixels, sizeof(unsigned char) * jpegSize);
   server_impl.setImageData(p);
